Made CameraPane a wrapper for layout views.
