+++ bgw-gui/src/main/kotlin/tools/aqua/bgw/builder/NodeBuilder.kt
@@ -17,7 +17,6 @@
 
 package tools.aqua.bgw.builder
 
-import javafx.scene.control.ScrollPane
 import javafx.scene.layout.Region
 import javafx.scene.layout.StackPane
 import tools.aqua.bgw.builder.DragDropBuilder.registerDragEvents
@@ -30,6 +29,7 @@ import tools.aqua.bgw.components.DynamicComponentView
 import tools.aqua.bgw.components.StaticComponentView
 import tools.aqua.bgw.components.container.GameComponentContainer
 import tools.aqua.bgw.components.gamecomponentviews.GameComponentView
+import tools.aqua.bgw.components.layoutviews.CameraPane
 import tools.aqua.bgw.components.layoutviews.LayoutView
 import tools.aqua.bgw.components.layoutviews.Pane
 import tools.aqua.bgw.components.uicomponents.UIComponent
@@ -49,6 +49,7 @@ object NodeBuilder {
           is GameComponentView -> ComponentNodeBuilder.buildGameComponent(componentView)
           is LayoutView<out ComponentView> ->
               LayoutNodeBuilder.buildLayoutView(scene, componentView)
+          is CameraPane<out LayoutView<*>> -> CameraPaneBuilder.buildCameraPane(scene, componentView)
           is UIComponent -> UINodeBuilder.buildUIComponent(componentView)
           is StaticComponentView<*> ->
               throw IllegalInheritanceException(componentView, StaticComponentView::class.java)
